dashboard/requirements.txt
streamlit
pandas
numpy
matplotlib
plotly
psycopg2-binary
python-dotenv
sqlalchemy
folium
streamlit_folium
